apps/subnet-calculator/apim-simulator/pyproject.toml
[project]
name = "apim-simulator"
version = "0.1.0"
description = "Config-driven Azure API Management simulator for local Docker-first testing"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
    "pyjwt[crypto]>=2.10.1",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "opentelemetry-exporter-otlp-proto-http>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-httpx>=0.61b0",
    "opentelemetry-instrumentation-logging>=0.61b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "integration: requires external services (e.g. Keycloak) running",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
